Move all DTOs to models directory
Moving frontend DTOs to their own files.
