+++ src/frontend/src/models/Attendance.ts
@@ -1,0 +1,9 @@
+interface Attendance {
+  key: number;
+  jobId: string;
+  jobName: string;
+  company: string;
+  date: string;
+}
+
+export default Attendance;
