+++ src/frontend/src/models/CheckInHistory.ts
@@ -1,0 +1,6 @@
+const checkInHistory = [
+  { key: "1", jobId: "J001", employeeId: "E001", employeeName: "John Doe", timeIn: "2025-03-20 08:00" },
+  { key: "2", jobId: "J002", employeeId: "E002", employeeName: "Jane Smith", timeIn: "2025-03-20 08:30" },
+];
+
+export default checkInHistory;
