+++ src/frontend/src/models/CheckOutHistory.ts
@@ -1,0 +1,6 @@
+const checkOutHistory = [
+  { key: "1", jobId: "J001", employeeId: "E001", employeeName: "John Doe", timeOut: "2025-03-20 08:00" },
+  { key: "2", jobId: "J002", employeeId: "E002", employeeName: "Jane Smith", timeOut: "2025-03-20 08:30" },
+];
+
+export default checkOutHistory;
