+++ src/frontend/src/models/Employee.ts
@@ -1,0 +1,8 @@
+interface Employee {
+  key: number;
+  id: string;
+  name: string;
+  contact: string;
+}
+
+export default Employee;
