+++ src/frontend/src/models/Job.ts
@@ -1,0 +1,10 @@
+interface Job {
+  key: string;
+  jobId: string;
+  purchaseOrderNo: string;
+  jobName: string;
+  companyName: string;
+  checked: boolean;
+}
+
+export default Job;
