+++ src/frontend/src/models/JobData.ts
@@ -1,0 +1,10 @@
+interface JobData {
+  key: string;
+  jobId: string;
+  purchaseOrder: string;
+  jobName: string;
+  requestor: string;
+  contact: string;
+}
+
+export default JobData;
