+++ src/frontend/src/models/Timesheet.ts
@@ -1,0 +1,10 @@
+interface Timesheet {
+  key: string;
+  date: string;
+  employeeId: string;
+  timeIn: string;
+  timeOut: string;
+  hoursWorked: string;
+}
+
+export default Timesheet;
