+++ src/frontend/src/models/Worker.ts
@@ -1,0 +1,5 @@
+export interface Worker {
+  key: string;
+  name: string;
+  role: string;
+}
